partially fixed
